Add dynamic boxHeader
